move scan over to lax_control_flow.py
